For the v1.2.4 release
- Adds support for Python 3.9 and 3.10
- Addresses deprecation warnings and other minor bug fixes
- Resolves dependency issues and includes an updated setup.py file
- Adds GitHub Actions workflows that trigger automatic tests on the latest Ubuntu OS, macOS, and Windows for Python versions 3.7 to 3.10 each time a push or pull request is made to the main branch
- An additional Jupyter Notebook example that demonstrates the core functionalities of MIDASpy

+++ MIDASpy/midas_base.py
@@ -1992,7 +1992,7 @@ def cat_conv(cat_data):
 
     for column in cat_data.columns:
         na_temp = cat_data[column].isnull()
-        temp = pd.get_dummies(cat_data[column], prefix=column)
+        temp = pd.get_dummies(cat_data[column], prefix=column, dtype=np.uint8)
         temp[na_temp] = np.nan
         cat_construct.append(temp)
         cat_col_names.append(list(temp.columns.values))
